requirements.txt
Django==5.1.4
django-cors-headers==4.3.1
djangorestframework==3.14.0
whitenoise==6.6.0
django-extensions==3.2.3
Pillow==10.1.0
qrcode==7.4.2
pywin32==306; sys_platform == "win32" 